Sign jar iff property jcecertAlias, else just copy unsigned to build/lib.
This commit enables opportunistic jarsigning (for JDKs which require JCA providers to be signed).
It does this by having two different (final) build steps depending on the presence of the jcecertAlias property.
If the property is not present, then the build jar is simply copied from the working directory into build/lib.
If the property is present, then it uses it (and other properties) to sign the jar in the working directory and
output the resulting jar into build/lib. This will let standard builds complete and allow our publishing systems
to change the behavior by having a gradle.properties file in an appropriate location.