Support testing remote jars and add primary artifact.
This adds support for testing jars in remote (staging) maven repositories to
make it easier for us to validate builds prior to release. It also adds an
empty JAR file as a primary target (classifier=null) which won't work for
any consumers but avoids confusing systems which expect /something/ to
be there.