Normalize logging.
Decrease logging so that there is no output under standard circumstances.
This includes either completely successful or completely unsuccessful loading.
Ensure that logging of failed self-tests remains high as this is still an
important issue. This is in preperation for #37 when logging will be both
more extensive and better controllable.