Add check for broken Ryzen RDRAND
Some AMD Ryzen CPUs get into a bad state where
the output of RDRAND becomes "stuck" on a value
of all 1s (e.g., every bit in the output is a one).
Unfortunately, it doesn't correctly set the carry flag
to indicate an RNG failure.

This change detects this case and handles it as a
standard RNG error.