Add check for broken Ryzen RDRAND (#67)
* Add check for broken Ryzen RDRAND

Some AMD Ryzen CPUs get into a bad state where
the output of RDRAND becomes "stuck" on a value
of all 1s (e.g., every bit in the output is a one).
Unfortunately, it doesn't correctly set the carry flag
to indicate an RNG failure.

This change detects this case and handles it as a
standard RNG error.

* Move CHANGELOG info back to 1.2.0

* Move stuck RDRAND to rng_retry_rdrand

This moves where we check for RDRAND being stuck to a better
location and also adds tests for this logic. These tests only
run with the 'coverage' target. I have validated that the
tests work by backing out the fixes and manually verifying
the failure.

* Indentation fix

* Just discard suspicious RDRAND values with no fixup

Also re-add new-line in test_rdrand.cpp

* Typo fix