Fix EC Curve selection by size
Now, when a new EC key pair is generated by size (rather than by curve), always select a 'secp*r1' curve.
For the standard sizes of 192, 224, 256, 384, and 521 this always matches the behavior of the SunEC provider.

This changes the behavior for sizes 192 and 256 from the secp*k1 to the secp*r1 curves (NIST P-192 and P-256) respectively.