Allow resources to be located while in jmod form
Some of the getResourceAsStream calls will fail if running as a jmod
integrated in a custom jdk. This is due to resource file encapsulation
blocking access using jar absolute paths for non open modules.

By always using paths relative to the Loader.class we can ensure the
code works as a jar or as a jmod.

Resource files within the jar have been moved to package:
com.amazon.corretto.crypto.provider.testdata