Drop DH-flavored KeyAgreement
Diffie-Hellman EVP-based key agreement [tests][1] are failing. it looks
like AWS-LC [supports ASN.1 DH][2], but not currently through the [EVP
API][3] (nor does [boringssl][4]). there was an [issue in aws-lc-old][5]
around this, but it doesn't appear to have been implemented (the ACCP
issue linked there looks related, but not quite the same as [5]).

This leaves us with (more or less) 3 options:

  1. implement EVP support for DH in AWS-LC
  2. implement DH-flavored KeyAgreement using lower-level APIs (this
     also  means we won't be sharing common logic with ECDH on the java
     side anymore)
  3. drop support for DH-flavored KeyAgreement entirely

Pending discussion with the team about FF DH's desirability, this commit
opts for option 3 as a first cut.

[1]: https://github.com/corretto/amazon-corretto-crypto-provider/blob/a72c6de38e108ceb34b5dcbbc31307a292e66122/tst/com/amazon/corretto/crypto/provider/test/EvpKeyAgreementTest.java#L81-L83
[2]: https://github.com/awslabs/aws-lc/blob/main/crypto/dh_extra/dh_asn1.c
[3]: https://github.com/awslabs/aws-lc/blob/main/crypto/evp_extra/p_methods.c#L15
[4]: https://github.com/google/boringssl/blob/master/crypto/evp/evp_asn1.c#L72
[5]: https://github.com/awslabs/aws-lc-old/issues/92