Allow SignatureException on ECDSA verify failure
This commit modifies the EvpSignatureTest to allow for
`SignatureException`s to be thrown upon ECDSA verification failure (as
opposed to simply returning `false`). Before this change, a number of
tests failed due to `SignatureException`s being thrown on ECDSA
verification failure (but not RSA verification†).

A relevant difference in behavior between OpenSSL and AWS-LC is that
OpenSSL [**does not** set][1] an error code when it fails to verify an
otherwise valid ECDSA signature while [AWS-LC **does**][2] (i.e. the
signature is well-formed and uncorrupted, but it is inauthentic per the
specified public key). Interestingly, AWS-LC's ancestor BoringSSL
[doesn't set an error][4] on ECDSA verification failure either.
AWS-LC's setting of a `BAD_SIGNATURE` error for _DSA_ verification
feailure is consistent with [AWS-LC's treatment of _RSA_ signature
verification failres][3].

When we (ACCP) see an error after attempting (non-RSA) verification, we
[throw an exception][8]. Unfortunately, because AWS-LC currently
conflates _corrupted_ and _inauthentic_ signatures under the same
`BAD_SIGNATURE` error code, we can not differentiate the cases, and thus
will never return `false` on ECDSA verification -- we'll either return
`true` if the signature is valid and authentic, else throw. The
[javadocs][9] for `Signature.verify` seem to indicate that we should
indeed return boolean `false` on verification failure. In order to do
this, we'll need to make changes to AWS-LC to either set no error code
upon verification failure or to set a distinct code which we can key off
of.

---

† A [prior commit][5] indicates that the JCA/JCE specification
stipulates that we suppress exceptions for RSA, but I was unable to find
any such stipulation in the [latest JCA docs][6] or [the javadoc for
`SignatureException`][7].

[1]: https://github.com/openssl/openssl/blob/5cbd2ea3f94aa8adec9b4486ac757d4d688e3f8c/crypto/ec/ecdsa_ossl.c#L458
[2]: https://github.com/awslabs/aws-lc/blob/main/crypto/fipsmodule/ecdsa/ecdsa.c#L194-L197
[3]: https://github.com/awslabs/aws-lc/blob/main/crypto/fipsmodule/rsa/rsa.c#L620-L621
[4]: https://boringssl.googlesource.com/boringssl/+/2214/crypto/ecdsa/ecdsa.c#215
[5]: https://github.com/corretto/amazon-corretto-crypto-provider/commit/dcbf8a7dc6fb5064b979739e2ffc590de61c39a2#diff-7fe3b81789f62d5bbf2ff0d8a6a8db06393630449ff914f9127f788084b06e49R481-R484
[6]: https://docs.oracle.com/en/java/javase/17/security/java-cryptography-architecture-jca-reference-guide.html
[7]: https://docs.oracle.com/javase/7/docs/api/java/security/SignatureException.html
[8]: https://github.com/corretto/amazon-corretto-crypto-provider/blob/develop/csrc/sign.cpp#L483
[9]: https://docs.oracle.com/javase/7/docs/api/java/security/Signature.html#verify(byte[])