Track AWS-LC as a git submodule
This change should allow for easier co-development of AWS-LC when
embedded in the ACCP source tree, as well as allowing us to track more
granular commits/rollbacks than if we were downloading releases.

To simplify the gradle build script, we require the user to recursively
populate the submodlue with AWS-LC's code. We throw an instructive error
message if this has not been done, and provide README documentation for
how to do so.