Support PSS padding for RSA signatures
This commit adds a [new flavor of RSA Signatures][1]: `RSASSA-PSS`.
PSS-padded RSA is described in [RFC 8017 section 8.1][2], with relevant
input parameter validation explained in [RFC 3447 section 9.1.1][3]. As
with OAEP RSA cipher padding, we mostly rely on AWS-LC to validate
message digest support, and explicity forbid anything outside of the SHA
family.

PSS allows users to specify the length (in bytes) of the randomly
generated salt. This value may not be less than 0 (JCE doesn't allow
negative salt lengths at all, and AWS-LC reserved them as special flag
values), so we enforce a minimum salt length of 0. The maximum is
specified in [RFC 3447 section 9.1.1][3], but relies on the key modulus
size to determine the maximum length. Since we don't always know the
key size when setParameter is called (key is specified on `init*`, and we
don't enforce ordering between calls to `setParameter` and `init*`), we
assume 2048 bit field size if no key is present. 2048 was selected
because it is the smallest field size that is still considered "secure",
and if a caller ends up using a smaller key, AWS-LC will detect this and
throw an error.

We switch the JCE provider for RSASSA-PSS parameterized tests in
EvpSignatureTest from BouncyCastle (BC) to [the SunRsaSign provider][6],
primarily due to differences in how they handle PSS parameters.
SunRsaSign [allows PSS message digest algorithm and MGF digest
algorithms to differ][4], while BC [does not][5]. So, in order to test
all possible combinations of PSS and MGF digest algorithms, we need to
migrate those tests to use SuncJCE as the JCE provider.

To increase compatibility accross providers, we allow callers to specify
`null` params in calls to `setParameter` (this is [consistent with
BC][7], but [not with SunRsaSign][8]. Similarly, we don't require
`setParameter` to be called on `RSASSA-PSS` instances before `init*` is
called, defaulting to digest algorithms of SHA1 unless otherwise
specified.

One point where we do diverge from SunRsaSign (but _not_ from BC) is
that we throw if callers attempt to update PSS parameters while
building or verifying a signature. We forbid this behavior to avoid
signature corruption due to digest algorithms being changed
while the signature/digest is still being updated.

Finally, @jakemas noted [a criticism][9] of PSS described by Peter
Gutman, the crux of which is:

> Now let's look at what an attacker can do with the above.  The
> immediately obvious, trivial attack is a hash-substitution attack.
> Unlike PKCS #1, PSS doesn't encode the hash algorithm used anywhere in
> the signature. [...] By changing hashAlgorithm to whatever you like, for
> example an algorithm of the same size that you know how to break, you
> can forge the signature.  Change the hash from SHA-2/256 to Streebog,
> Blake2, Keccak, CRC-256, XOR256, whatever you like, PSS won't detect the
> change, making it only as strong as the weakest hash algorithm of the
> same bit width that the victim will accept.

Unfortunately, because PSS is already standardized, there's not much we
can do to mitigate this in our impelementation. The main mitigation we
employ is to explicitly disallow non-SHA hash functions (e.g. MD5). All
SHA-family hash functions are currently believed to be secure for use in
PSS (indeed, the oldest, SHA1, is the default for Java) so no immediate
threat remains. If SHA1 is deemed insecure for use in PSS at some point
in the future, however, we will need to restrict its specification as a
PSS parameter.

[1]: https://docs.oracle.com/en/java/javase/11/docs/specs/security/standard-names.html#signature-algorithms
[2]: https://datatracker.ietf.org/doc/html/rfc8017#section-8.1
[3]: https://datatracker.ietf.org/doc/html/rfc3447#section-9.1.1
[4]: https://github.com/corretto/corretto-11/blob/f2c5b39cdb2afc9a9101620d14cf34c7448f826d/src/java.base/share/classes/sun/security/rsa/RSAPSSSignature.java#L290-L333
[5]: https://github.com/bcgit/bc-java/blob/bc3b92f1f0e78b82e2584c5fb4b226a13e7f8b3b/prov/src/main/java/org/bouncycastle/jcajce/provider/asymmetric/rsa/PSSSignatureSpi.java#L239-L242
[6]: https://docs.oracle.com/javase/7/docs/technotes/guides/security/SunProviders.html#SunRsaSignProvider
[7]: https://github.com/bcgit/bc-java/blob/bc3b92f1f0e78b82e2584c5fb4b226a13e7f8b3b/prov/src/main/java/org/bouncycastle/jcajce/provider/asymmetric/rsa/PSSSignatureSpi.java#L199-L207
[8]: https://github.com/corretto/corretto-11/blob/f2c5b39cdb2afc9a9101620d14cf34c7448f826d/src/java.base/share/classes/sun/security/rsa/RSAPSSSignature.java#L292-L294
[9]: https://marc.info/?l=cryptography&m=157345122116408&w=2