Enable FIPS builds with -DFIPS=true
Initial commit to allow ACCP to use the FIPS build of AWS-LC.
This requires changing our linking from static to dynamic and
introducing a new shared object specifically to load the library.

We also introduce the `PRESERVE_NATIVE_LIBRARIES` debug flag to
make it easier to debug running instances of ACCP by not deleting
the temporary files containing libraries until the JVM terminates.