Split CI into different jdk versions (#216)
This PR spits the JDK tests across different dimensions
according to its version. The TEST_JAVA_HOME variable is added
to each build to specify the path to the java version we want
to test upon. ACCP still needs a JDK version > 10 for building,
and JDK versions lower than that are only used for testing.