Fix libaccp builds for GCC 4.1.2 (#225)
This commit fixes builds for ACCP's native library on GCC 4.1.2, which
only supports a subset of C++11. This issue was first discovered while
trying to build ACCP internally on RHEL5, which failed. Two of the
changes in this commit are trivial syntax/const semantics changes.

Another change accounts for C++ not having added [`std::move` to the
standard library until C++11][1], and also fixes a subtle "bug" that was
introcude [here][2] (prompted by this][3]). The change to using
`BigNumObj::move` instead of `std::move` under C++11 shouldn't make a
significant difference -- the two behave the same in that they both give
ownership the passed BN to the calling instance. In C++ versions less
than 11, however, `std::move` doesn't exist so the code doesn't compile.

Finally, we update our AWS-LC dependency commit to incorporate [a recent
change][2] fixing AWS-LC C++ builds on GCC 4.1.2.

[1]: https://en.cppreference.com/w/cpp/utility/move
[2]: https://github.com/corretto/amazon-corretto-crypto-provider/pull/132/commits/b285ff31cf5474b6834d43c870455cfcfffbbd50#diff-3c12dad8524f0f17ae8b278d84f58fa405409117925d51bb9d08fc9758ae1382
[3]: https://github.com/corretto/amazon-corretto-crypto-provider/pull/132#discussion_r861967037
[4]: https://github.com/awslabs/aws-lc/commit/42a8dbacd473d2ad95f9d0d5e47ba95a4e92f4c3