Implement AES KWP
This commit adds a new CipherSpi implementation, AesKeyWrapSpi. This
cipher can be instantiated as AES with mode KWP, and NoPadding. AWS-LC
only provides one-shot functions for key wrapping and unwrapping, so we
buffer any incremental data fed to the Cipher until finalization, at
which point we pass all buffered data over the JNI and call AWS-LC. AES
KWP is defined in [RFC-5649][1], which we rely on for output buffer
sizing and other implementation-adjacent details/input validation.

While JCE and BouncyCastle support callers specifying custom IVs, AWS-LC
does not so neither do we. We do not throw when a caller tries to
set IV's via `setParameters` or similar, we instead ignore the call and
return `null` on calls to `getParameters` and `getIv`.

As described in [this document][2], NIST provides Known Answer Tests for
AES KWP. We pull down [these test vectors][3] and include `gzip`'s of
them in the repo. We then add AesKwpKatTest to read these test cases in,
and execute them against ACCP. This commit follows prior art for KAT
tests in AesGcmKatTest. We also add many parameterized and input
validation tests, achieving 87% line coverage and 80% branch coverage.

Finally, we fix up CMakeLists.txt to include the AES GCM KAT tests in
its test builds/runs, which had previously been dropped from the test
set.

[1]: https://datatracker.ietf.org/doc/html/rfc5649
[2]: https://csrc.nist.gov/CSRC/media/Projects/Cryptographic-Algorithm-Validation-Program/documents/mac/KWVS.pdf
[4]: https://csrc.nist.gov/CSRC/media/Projects/Cryptographic-Algorithm-Validation-Program/documents/mac/kwtestvectors.zip