Report coverage in CodeBuild
This commit configures coverage reports for CodeBuild, as documented
[here][1]. This is very straightforward for our Jacoco java coverage
reports, as CodeBuild natively supports that report format. For our gcov
C++ coverage reports for libaccp, however, we need to install and use
[`gcovr`][2] to convert gcov reports to a compatible format. Coverage is
only reported for our basic tests (either via the `release` target or,
in the case of Java8, explicitly specifying the `coverage` target
alongside `test`).

CodeCommit's coverage reports should be useful for PR reviewers, but
they do not provide granular, line-by-line coverage information or a
diff overlay view. If we want something more granular, we may need to
evaluate 3rd party tools [like codecov][3].

This commit also addresses [Issue #239][4], using CMake's [`file(GLOB`
function][5] to find relevant src and test files, automatically
generating the file of files to include java in builds. This eliminates
the need to manually enumerate our src/tst files in CMakeLists.txt. We
use the `CONFIGURE_DEPENDS` flag, instructing CMake to watch for new (or
removed) files in those directories, and regenerate the build system if
it detects a change. However, this feature is not available until 3.12,
so we need to condition its use on CMake version (per @sacchaing, RHEL5
only supports up to CMake 3.9).

[1]: https://docs.aws.amazon.com/codebuild/latest/userguide/code-coverage-report.html
[2]: https://gcovr.com/en/stable/
[3]: https://docs.codecov.com/docs/github-checks
[4]: https://github.com/corretto/amazon-corretto-crypto-provider/issues/239
[5]: https://cmake.org/cmake/help/latest/command/file.html#glob