Fix gcovr reports on legacy ubuntu 10 build
Ubuntu 10 is very outdated, so we have to work around some constraints.
Firstly, while python3.1 is available on ubuntu 10 via `apt-get`, pip3
is not. So, we need to use legacy pip. Secondly, that legacy pip can't
install packages over the network so we need to download the tarball as
a copied dependency and install it from the filesystem (as we do with
other packages in that build). Finally, only gcovr 2.0 will build on
our legacy ubuntu image (we use gcovr 5.x elsewhere), so we need to
switch to using the `--xml` output configuration flag. Thankfully,
`--cobertura` in later versions is just an alias to `--xml`, so we can
use the same flag across versions.