Support MacOS
This commit adds functional and CI support for MacOS. The main blocker
on supporting MacOS was a runtime linking error caused by libaccp trying
to load a `libcrypto` it found in some `/usr` system directory on MacOS,
instead of the temporary directory that ACCP creates and restricts
`rpath` to during linking. The reason for this is that the `rpath` value
of `$ORIGIN` is specific to linux's linker, and is meaningless on MacOS.
So, the MacOS linker consulted a non-existent `rpath`, didn't find
anything, and fell back to system library directories. The fix is to
detect OS type at build time and use the appropriate value for `rpath`
(on MacOS that's `@loader_path`, see the "Locating External Resources"
section of [this documentation page][1]).  We also explicity set
`CMAKE_MACOSX_RPATH` on MacOS, which is documented [here][2].

To prevent regressions expanding RPATH beyond one entry, we add a
test check that asserts that libaccp only has a single RPATH entry
that's set to the appropriate value on all supported platforms.

The other pieces of this commit are related to CI. In order for the
coverate steps to pass, we need to install `lcov` and `gcovr`
dependencies on the GitHub Actions MacOS CI images. Lastly, now that
this runs on Mac the Apple JCE providers in our pre-baked properties
files are actually used (they're ignored on Linux, where they're
unavailable). This resulted in the same provider being installed twice,
causing SecurityPropertiesTester failures. We resolve this issue by
removing the class-specified entry in favor of the name-specified entry.
Because the [AppleProvider][3] is [not exported][4] from its module in
JDK 11, we cannot use the fully qualified class specification (see
comment in `etc/amazon-corretto-crypto-provider.security`) on JDK 9+. By
using only the name specification, there may be some issues on JDK8 on
Mac when using this providers file, but the impact should be minor -- no
regressions (because we didn't support MacOS before). Users are
encouraged to upgrade to more moder JDKs or manually modify/override the
supplied properties file.

This commit resolves [Issue #41][5].

[1]: https://developer.apple.com/library/archive/documentation/DeveloperTools/Conceptual/DynamicLibraries/100-Articles/DynamicLibraryDesignGuidelines.html#//apple_ref/doc/uid/TP40002013-SW21
[2]: https://cmake.org/cmake/help/latest/prop_tgt/MACOSX_RPATH.html
[3]: https://github.com/corretto/corretto-11/blob/develop/src/java.base/macosx/classes/apple/security/AppleProvider.java
[4]: https://github.com/corretto/corretto-11/blob/develop/src/java.base/macosx/classes/module-info.java.extra
[5]: https://github.com/corretto/amazon-corretto-crypto-provider/issues/41