Fix 521-bit EC tests for Java10, preserve equality where we can
A more thorough summary of the issue will be provided before this PR is
promoted to Draft status, but the long and the short of it is that
Java10 has a bug in it that predominantly affects 521-bit EC private
keys. The bug is [here][1], and results in the truncation and
mis-encoding of EC keys who doen't have the most significant (i.e.
"leftmost" in big-endian) bit set.

More analysis to come.

[1]: https://github.com/openjdk/jdk10/blob/master/src/jdk.crypto.ec/share/classes/sun/security/ec/ECPrivateKeyImpl.java#L94