Implement AlgorithmParameters for EC
This commit registers support for EC AlgorithmParameters with the JCA,
in an effort to get ACCP closer to functioning as a stand-alone
providers. At startup (i.e. `static` initialization time), we query
AWS-LC for which curves are avaialable at startup and cache the relevant
information in order to avoid redundant JNI calls fetching the same,
invariant information.

For some tests to function with correctly with the new
AlgorithmParameters on Java17+, we need to add the sun.security.util
package (internal to JDK, not exported outside of the java.base module)
to the list of specified "opens", allowing for deep reflection.  We
attempt reflection in some cases across java versions because JCE will
sometimes pass `ECKeySizeParameterSpec` instances to
`AlgorithmParameters.init`, but `ECKeySizeParameterSpec` is not a public
class in JDK8 or exported within a module in JDK11+. So, as a "best
effort", we try to handle this using reflection to extract the key size,
which must be explicitly allowed for JDK-internal classes/modules in
Java17+. It's not an ideal, but is an unfortunate reality of the JDK's
module configuration and interface limitations.

I confirmed that the new EcParameters class has full coverage, save for
a few exception-throwing cases that are difficult to trigger due to
current inability to mock our JNI/AWS-LC layer.