Overhaul benchmarks, migrate to JMH (#263)
This commit rewrites our existing benchmarks and migrates them to
[JMH][1], integrated with our gradle build system using [a plugin][2].
We decided to migrate to JMH for a few reasons: reproducibility,
isolation, accounting for JVM JIT optimizations, and (not least of
which!) pretty graphs and configurable output format.  For more details
on the tradeoffs between conventional benchmarks and JMH, see [this blog
post][3].

Other than the benchmarks themselves, the real meat of this change lives
in `build.gradle`. The JMH plugin introduced a few gradle tasks whose
names collided with some of our existing tasks, so some existing task
names were made more specific. The tricky parts of this change were
modifying the JMH task sourceSets and JVM args to point to all the
relevant source files in our codebase. Similarly, we had to add some
static resources (as symlinks to avoid duplication) to a JMH resources
directory where gradle can find them and bundle them into the benchmark
JAR.

We also provide a `./bin/run-benchmarks.sh` script to install
dependencies (if needed), correctly invoke the JMH benchmarks, and
generate a report.

The benchmarks currently compare ACCP against BouncyCastle and built-in
Java providers. If we want to add BouncyCastle FIPS to the mix, it
should be easy to add [`BCFIPS` as a maven dependency][4] for `jmh` as
we did for vanilla BouncyCastle then include it in our providers lists
in the individual benchmarks.

These JMH changes were backported to ACCP 1.6 on [the benchmarking-1.x
branch][5] for historical comparison. You can see what the benchmark
reports look like here:

- [ACCP 1.6 on c6g.large (Graviton 2)][6]
- [ACCP 1.6 on c7g.large (Graviton 3)][7]
- [ACCP 2.0 on c6g.large (Graviton 2)][8]
- [ACCP 2.0 on c7g.large (Graviton 3)][9]

[1]: https://github.com/openjdk/jmh
[2]: https://github.com/melix/jmh-gradle-plugin
[3]: https://www.oracle.com/technical-resources/articles/java/architect-benchmarking.html
[4]: https://mvnrepository.com/artifact/org.bouncycastle/bc-fips
[5]: https://github.com/corretto/amazon-corretto-crypto-provider/tree/benchmarking-1.x
[6]: https://dr3qfay5p537.cloudfront.net/accp-1-graviton-2/jmh/index.html
[7]: https://dr3qfay5p537.cloudfront.net/accp-1-graviton-3/jmh/index.html
[8]: https://dr3qfay5p537.cloudfront.net/accp-2-graviton-2/jmh/index.html
[9]: https://dr3qfay5p537.cloudfront.net/accp-2-graviton-3/jmh/index.html