Only register LibCryptoRng by default in non-FIPS mode
See the updated README for this option's motivation.

Unlike `registerEcParams`, we only provide this option as a JVM property
and not as a runtime method call. The reason for this is that the JCA
treats SecureRandomly differently from other algorithms, maintining
provider registry state across both [SecureRandom][1] and [Provider][2]
classes. This significantly complicates registering SecureRandom
implementations post-initialization for a given provider (e.g. ACCP)
after instances of SecureRandom have already been generated from a
different provider (e.g. SUN). So, for now, we only expose this option
as a JVM property set at process start.

[1]: https://github.com/corretto/corretto-11/blob/1fee80d48d98e91a6a7ad1ca24220ccc7c148ef3/src/java.base/share/classes/java/security/SecureRandom.java#L274
[2]: https://github.com/corretto/corretto-11/blob/1fee80d48d98e91a6a7ad1ca24220ccc7c148ef3/src/java.base/share/classes/java/security/Provider.java#L1411