SecretKeyGenerator's DefaultSecureRandomSupplier always uses LibCryptoRng
We recently made [a change][1] to ACCP making the JCA registry of a
SecureRandom implementation opt-in in FIPS mode. This had the unintended
effect of breaking SecretKey generation in FIPS mode, as our
SecretKeyGenerator internally sourced its LibCryptoRng instance from the
JCA. Because we didn't register LibCryptoRng with the JCA by default in
FIPS mode, exceptions got thrown.

The fix here is to just instantiate LibCryptoRng directy instead of
obtaining it through the JCA. This has the added benefit of making sure
we're _always_ using AWS-LC's DRBG for SecretKey generation by default,
eliminating the possiblity of throw/fallback flows.

[1]: https://github.com/corretto/amazon-corretto-crypto-provider/pull/286