Directly buffer ciphertext bytes during doUpdate() for AES-GCM decryption
Our implementation of engineUpdate for AES-GCM does not return intermediate data to the output buffer. Instead, it simply caches the data to an internal buffer until doFinal() is called. This is to ensure that we validate the tag before returning the plain text and not leave the caller holding the bag.

That said, ideally, doUpdate should simply perform a memcopy into the internal buffer and return. If necessary, it should perform a dynamic re-allocation to accommodate additional data (amortized O(n). This change implements that ideal by completely bypassing the complicated workflow we were doing previously.

Things we avoid:
- Creating a complete copy of the input buffer which is done to avoid overwriting the input. However, in our decrypt implementation we do not write any output so this was pointless. O(n) memory savings.
- Tight loop which chunked our loading of the internal decrypt buffer in 64k increments.
  + O(n) Allocations of small `ShimArray` objects on each loop iteration.
  + O(n^2) memory I/O and O(n) additional memory due to the array resizing from the incremental loading of data

The end result of this is that we will have much lower garbage collector pressure overall for larger payload sizes (>10MB). Based on the benchmarks provided in issue #271, this is a 33% improvement. It's still not where we want to be but it gets us closer.