Fix bug in output buffer size check
This is calling the wrong function and getting the wrong size. Note how
the exception's error message used a different function than the one
used in the error checking logic. For encryptions, this causes our
safety check to be off by the tag length. Currently unused for
decryptions, but this would return 0 when it should be the size of the
plaintext.