Optimize handling of overlapping input/output buffers for AES-GCM
Currently, ACCP does not tolerate any overlap between input/output buffers. If it detects any overlap, it will buffer a copy of the entire input prior to proceeding (consuming O(n) more memory). This is to prevent any risk of the output of the operation from overwriting the input prior to processing that input. However, it is a common case for our clients (e.g. Sun JSSE's TLS implementation) to preserve memory by reusing the input/output buffers since they have little use for the input once it has been successfully encrypted/decrypted.

AES, as a block cipher, operates on blocks. It reads one block of input into intermediate memory for processing and then writes one block of output. Implementations of AES can tolerate writing the output to the exact same block of memory where it read the input. Where we run into trouble is whenever the output block overwrites data from any input block beyond the block that was just processed.

This changeset relaxes the boundary checking in the input/output arrays so that it allows the case where we overlap exactly or any case where the output lags behind the input. We still safely handle cases where the output overwrites the input by detecting and cloning the input buffer as necessary. We also check for cases where the output leads the input, but beyond the limit of the input buffer and allow this configuration without cloning memory. The end result is that we unlock another 30% improvement in memory consumption under the benchmarks in issue #271.