Invert function overloading order for AES-GCM doFinal decryption
There are 3 overloaded methods for calling into a Cipher module which implement different ways of returning the output of the operation. From a memory management perspective these can be differentiated by whether or not the caller is responsible for allocating memory for the output.

Previously in our implementation, the top layer of the call stack is a method which allocates a buffer for the output and returns the result. For other interfaces in which the caller is responsible for allocating the output buffer we would allocate yet another buffer and copy it over to the buffer the caller gave us. This results in O(n) additional memory allocations and I/O.

This change inverts the call stack so that the top layer returns the result into an output buffer passed by the caller. That way, we only allocate a new output buffer when one is not provided. This gives us another 33% improvement in memory allocation from the benchmarks provided in issue #271.